huggingface-space/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
llama-cpp-python==0.2.90
huggingface-hub==0.20.3
